requirements.txt
Django
django-s3-sqlite
django-environ
zappa
virtualenv
django-s3-storage
sentry_sdk
zappa_sentry
django-cors-headers
pyyaml
uritemplate
markdown
pygments
django-filter
django-guardian
djangorestframework